server code to integrate Elasticsearch for document indexing and searching, replacing Pinecone. Update CORS configuration and document handling logic. enhanced AI Capabilities
