+++ api/server.js
@@ -1,340 +1,412 @@
 const path = require('path');
-require('dotenv').config({ path: path.join(__dirname, '.env') });
+require('dotenv').config({ path: path.join(__dirname, '.env', '.env-local') });
 const express = require('express');
 const bodyParser = require('body-parser');
-const { Pinecone } = require('@pinecone-database/pinecone');
+const { Client } = require('@elastic/elasticsearch');
 const { OpenAI } = require('openai');
 const fs = require('fs');
-const cors = require('cors');
+const cors = require("cors");
 const mammoth = require('mammoth');
 const chokidar = require('chokidar');
 
-// Pinecone client setup
-const pinecone = new Pinecone({ apiKey: process.env.PINECONE_API_KEY });
-const indexName = 'text-embedding-ada';
-const index = pinecone.Index(indexName);
+// Configurazione di Elasticsearch
+const esClient = new Client({ 
+    node: 'http://localhost:9200',
+    auth: {
+        username: process.env.ES_USERNAME,
+        password: process.env.ES_PASSWORD
+    }
+});
 
-// OpenAI client setup
-const openaiClient = new OpenAI({ apiKey: process.env.OPENAI_API_KEY });
+const openai = new OpenAI({
+    apiKey: process.env.OPENAI_API_KEY
+});
 
-// Express app setup
 const app = express();
 app.use(express.json());
-const port = process.env.PORT || 3000;
+const port = 3000;
 app.use(bodyParser.json());
 
-// CORS configuration
+//CORS//
 const corsOptions = {
-  origin: function (origin, callback) {
-    const allowedOrigins = [process.env.FE_DOMAIN, process.env.FE_DOMAIN_WWW, 'http://localhost:8080'];
-    if (!origin || allowedOrigins.includes(origin)) {
-      callback(null, true);
-    } else {
-      callback(new Error('Not allowed by CORS'));
+    origin: function (origin, callback) {
+      const allowedOrigins = [process.env.FE_DOMAIN, process.env.FE_DOMAIN_WWW];
+      if (!origin || allowedOrigins.includes(origin)) {
+        callback(null, true);
+      } else {
+        callback(new Error("Not allowed by CORS"));
+      }
+    },
+    methods: "GET,HEAD,PUT,PATCH,POST,DELETE",
+    credentials: true,
+    optionsSuccessStatus: 200,
+  };
+  
+  app.use(cors(corsOptions));
+
+const baseIndexName = 'myapp_index';
+
+let indexName = baseIndexName;//`${baseIndexName}_${Date.now()}`;
+
+// Creazione dell'indice (se non esiste)
+const createIndex = async () => {
+    try {
+        const exists = await esClient.indices.exists({ index: indexName });
+        if (!exists.body) {
+            await esClient.indices.create({
+                index: indexName,
+                body: {
+                    settings: {
+                        number_of_shards: 1,
+                        number_of_replicas: 0
+                    },
+                    mappings: {
+                        properties: {
+                            title: { type: 'text' },
+                            content: { type: 'text' },
+                        }
+                    }
+                }
+            });
+            console.log(`Indice "${indexName}" creato.`);
+        } else {
+            console.log(`Indice "${indexName}" già esistente.`);
+        }
+    } catch (error) {
+        console.error('Errore durante la creazione dell\'indice:', error);
     }
-  },
-  methods: 'GET,HEAD,PUT,PATCH,POST,DELETE',
-  credentials: true,
-  optionsSuccessStatus: 200,
-};
-app.use(cors(corsOptions));
-
-// Function to split text into larger chunks
-const splitIntoChunks = (text) => {
-  const paragraphs = text.split(/\n\n+/).filter(chunk => chunk.trim().length > 0);
-  const chunkSize = 500;
-  const chunks = [];
-  let currentChunk = '';
-
-  for (const paragraph of paragraphs) {
-    if (currentChunk.length + paragraph.length < chunkSize) {
-      currentChunk += (currentChunk ? '\n\n' : '') + paragraph;
-    } else {
-      if (currentChunk) chunks.push(currentChunk);
-      currentChunk = paragraph;
-    }
-  }
-  if (currentChunk) chunks.push(currentChunk);
-  return chunks;
 };
 
-// Function to generate embeddings using OpenAI
-const generateEmbedding = async (text) => {
-  try {
-    const response = await openaiClient.embeddings.create({
-      model: 'text-embedding-ada-002',
-      input: text,
-    });
-    return response.data[0].embedding;
-  } catch (error) {
-    console.error('Error generating embedding:', error);
-    throw error;
-  }
-};
-
-// Index a document by splitting into chunks and upserting to Pinecone
-const indexDocument = async (filePath) => {
-  try {
-    const { value: text } = await mammoth.extractRawText({ path: filePath });
-    const chunks = splitIntoChunks(text);
-    console.log(`Extracted ${chunks.length} chunks from "${filePath}"`);
-    if (chunks.length === 0) {
-      console.warn(`No chunks found in "${filePath}"`);
-      return;
-    }
-    const vectors = await Promise.all(
-      chunks.map(async (chunk, i) => {
-        const embedding = await generateEmbedding(chunk);
-        const id = `${Buffer.from(filePath).toString('base64')}_${i}`;
-        return {
-          id,
-          values: embedding,
-          metadata: {
-            file_path: filePath,
-            chunk_text: chunk,
-          },
-        };
-      })
-    );
-    await index.upsert(vectors);
-    console.log(`Document "${filePath}" indexed with ${chunks.length} chunks.`);
-  } catch (error) {
-    console.error(`Error indexing document "${filePath}":`, error);
-  }
-};
-
-// Delete vectors associated with a document
-const deleteDocumentVectors = async (filePath) => {
-  try {
-    await index.delete({ filter: { file_path: { $eq: filePath } } });
-    console.log(`Vectors for document "${filePath}" deleted.`);
-  } catch (error) {
-    console.error(`Error deleting vectors for document "${filePath}":`, error);
-  }
-};
-
-// Read documents from folder and index them
+// Funzione per leggere i documenti da una cartella
 const readDocumentsFromFolder = async (folderPath) => {
-  try {
     const files = fs.readdirSync(folderPath);
     const indexPromises = files.map(async (file) => {
-      const filePath = path.join(folderPath, file);
-      if (fs.statSync(filePath).isDirectory()) {
-        const currentDepth = folderPath.split(path.sep).length - baseFolderDepth + 1;
-        if (currentDepth < maxDepth) {
-          await readDocumentsFromFolder(filePath);
+        const filePath = path.join(folderPath, file);
+        if (path.extname(file) === '.docx') {
+            try {
+                const { value: content } = await mammoth.extractRawText({ path: filePath });
+                await indexDocument(file, content);
+            } catch (error) {
+                console.error(`Errore nella lettura del documento "${file}":`, error);
+            }
+        } else {
+            console.log(`File "${file}" non è un documento DOCX, verrà ignorato.`);
         }
-      } else if (path.extname(file).toLowerCase() === '.docx') {
-        await indexDocument(filePath);
-      } else {
-        console.log(`File "${file}" is not a DOCX document, skipping.`);
-      }
     });
+
     await Promise.all(indexPromises);
-  } catch (error) {
-    console.error(`Error reading folder ${folderPath}:`, error);
-  }
 };
 
-// Cache for recent search results
+// Funzione per indicizzare un documento con un id unico
+const indexDocument = async (title, content) => {
+    try {
+        // Usa il nome del file come id unico per prevenire duplicati
+        const id = title;
+
+        // Indica un documento specificando un id
+        await esClient.index({
+            index: indexName,
+            id,  // Questo sovrascriverà il documento se esiste già
+            body: {
+                title,
+                content,
+            },
+            op_type: 'index'  // 'index' sovrascrive o crea se non esiste
+        });
+
+        console.log(`Documento "${title}" indicizzato con successo.`);
+    } catch (error) {
+        console.error(`Errore durante l'indicizzazione del documento "${title}":`, error);
+    }
+};
+
+let lastNDocuments = [];
+let results = [];
+const searchDocuments = async (user_query, limit) => {
+    try {
+        // Get keywords from OpenAI
+        const aiResponse = await openai.chat.completions.create({
+            model: 'gpt-3.5-turbo',
+            messages: [
+                { "role": "system", "content": "You will be provided with a user query. Your goal is to extract a few keywords from the text to perform a search.\nKeep the search query to a few keywords that capture the user's intent.\nOnly output the keywords, without any additional text." },
+                { "role": 'user', content: user_query }
+            ],
+        });
+        const query = aiResponse.choices[0].message.content;
+        console.log("searchDocuments query", query);
+
+        // Search in Elasticsearch
+        const searchResponse = await esClient.search({
+        index: indexName,
+        body: {
+            query: {
+                bool: {
+                    must: [
+                        {
+                            multi_match: {
+                                query: query,
+                            fields: ['title', 'content'],
+                            type: 'most_fields',
+                            operator: 'and',  // Richiede che tutte le parole siano presenti
+                            fuzziness: '0'  // Disabilita il fuzziness per risultati più precisi
+                        }
+                        }
+                    ],
+                    should: [
+                        {
+                            match_phrase: {
+                                content: {
+                                    query: query,
+                                    slop: 0
+                                }
+                            }
+                        },
+                    ],
+                    minimum_should_match: 1
+                }
+            },
+            highlight: {
+                fields: {
+                    content: {
+                        fragment_size: 150,
+                        number_of_fragments: 3
+                    }
+                }
+            },
+            suggest: {
+                text: query,
+                simple_phrase: {
+                    phrase: {
+                        field: "content",
+                        size: 2,
+                        gram_size: 3,
+                        direct_generator: [{
+                            field: "content",
+                            suggest_mode: "always"
+                        }],
+                        highlight: {
+                            pre_tag: "<em>",
+                            post_tag: "</em>"
+                        }
+                    }
+                }
+            }
+        }
+    });
+
+        // Process search results
+        let results = searchResponse.hits.hits.map(hit => ({
+        id: hit._id,
+        title: hit._source.title,
+        highlight: hit.highlight.content
+    }));
+
+        lastNDocuments = searchResponse.hits.hits.slice(-limit).map(hit => hit._source.content);
+
+        // If no results, try with suggestions
+        if (results.length === 0 && searchResponse.suggest && searchResponse.suggest.simple_phrase[0].options.length > 0) {
+            const suggestedText = searchResponse.suggest.simple_phrase[0].options[0].text;
+        console.log("Nessun risultato trovato. Riprovo con suggerimento:", suggestedText);
+
+            // Search again with suggestion
+            const suggestedResponse = await esClient.search({
+            index: indexName,
+            body: {
+                query: {
+                    bool: {
+                        must: [
+                            {
+                                multi_match: {
+                                    query: suggestedText,
+                                    fields: ['title', 'content'],
+                                    type: 'most_fields',
+                                    operator: 'and',
+                                    fuzziness: 'AUTO'
+                                }
+                            }
+                        ],
+                        should: [
+                            {
+                                match_phrase: {
+                                    content: {
+                                        query: suggestedText,
+                                        slop: 2
+                                    }
+                                }
+                            },
+                        ],
+                        minimum_should_match: 1
+                    }
+                },
+                highlight: {
+                    fields: {
+                        content: {
+                            fragment_size: 150,
+                            number_of_fragments: 3
+                        }
+                    }
+                },
+            }
+        });
+        
+        // Aggiorna i risultati con quelli basati sul suggerimento
+            results = suggestedResponse.hits.hits.map(hit => ({
+            id: hit._id,
+            title: hit._source.title,
+            highlight: hit.highlight.content
+        }));
+        
+            lastNDocuments = suggestedResponse.hits.hits.slice(-limit).map(hit => hit._source.content);
+    }
+
+    return results;
+    } catch (error) {
+    console.log("searchDocuments error", error);
+    return [];
+}
+};
+
+
+app.post('/search_in_doc', async (req, res) => {  
+    const { query, limit, followUp, followUpText } = req.body;
+    
+    try {
+        if (!followUp) {
+            lastNDocuments = [];
+            results = [];
+            
+            results = await searchDocuments(query, limit);
+            
+            if (results.length > 0) {
+                const explanation = await generateGroupedExplanation(query, lastNDocuments);
+                res.json({ results, explanation });
+            }
+            else {
+                // Provide a more helpful message when no matches are found
+                const explanation = "Mi dispiace, non ho trovato corrispondenze nei documenti disponibili. Prova a riformulare la domanda in modo diverso o usa termini più generali.";
+                res.json({ results: [], explanation });
+            }
+
+        } else {
+            // Richiesta di follow-up
+            const explanation = await generateGroupedExplanation(followUpText, lastNDocuments, true, followUpText);
+            res.json({ results: [], explanation });
+        }
+    } catch (error) {
+        console.error('Errore nella ricerca:', error);
+        res.status(500).send('Errore nella ricerca');
+    }
+});
+
+// Variabili per cache temporanea
 let lastSearchPhrase = null;
 let lastExplanation = null;
-let lastNDocuments = [];
 
-// Search documents using Pinecone
-const searchDocuments = async (query, limit = 5) => {
-  try {
-    console.log('Searching for query:', query);
-    const queryEmbedding = await generateEmbedding(query);
-    console.log('Query embedding generated, length:', queryEmbedding.length);
-    const queryResponse = await index.query({
-      vector: queryEmbedding,
-      topK: limit,
-      includeMetadata: true,
-    });
-    console.log('Pinecone query response:', JSON.stringify(queryResponse, null, 2));
-    const results = queryResponse.matches.map((match) => ({
-      id: match.id,
-      score: match.score,
-      chunk_text: match.metadata.chunk_text,
-      file_path: match.metadata.file_path,
-    }));
-    console.log('Search results:', results);
-    lastNDocuments = results.map((result) => ({
-      title: path.basename(result.file_path),
-      content: result.chunk_text,
-    }));
-    console.log('Updated lastNDocuments:', lastNDocuments);
-    return {
-      results,
-      totalHits: queryResponse.matches.length,
-    };
-  } catch (error) {
-    console.error('Search error:', error);
-    return { results: [], totalHits: 0 };
-  }
+const generateGroupedExplanation = async (searchPhrase, documentContents, followUp,followUpText = '') => {
+    try {
+        let prompt;
+        const summarizedContents = documentContents.map(doc => doc.slice(0, 200));  // Usa solo i primi 200 caratteri di ogni documento
+        if (followUp && lastExplanation) {
+            // Prompt di follow-up con `followUpText`
+            prompt = `Hai chiesto chiarimenti riguardo a "${searchPhrase}".\n` +
+                     `Richiesta originale: "${lastSearchPhrase}"\n` +
+                     `Spiegazione precedente: "${lastExplanation.slice(0, 100)}""\n` +
+                     `Contenuti rilevanti dei documenti:\n` +
+                     `${summarizedContents.join('\n')}\n` +
+                     `Per favore, rispondi alla tua richiesta di chiarimento: "${followUpText}".`;
+        } else {
+            // Nuova richiesta di spiegazione
+            prompt = `Spiega "${searchPhrase}" in 30 parole:\n${summarizedContents.join('\n')}`;
+        }
+        
+        //console.log("generateGroupedExplanation prompt", prompt);
+
+        const response = await openai.chat.completions.create({
+            model: 'gpt-3.5-turbo',
+            messages: [{ role: 'user', content: prompt }],
+        });
+console.log("generateGroupedExplanation response", response);
+        const explanation = response.choices[0].message.content.trim();
+
+        // Aggiorna cache solo se è una nuova ricerca
+        if (!followUp) {
+            lastSearchPhrase = searchPhrase;
+            lastExplanation = explanation.length;
+        }
+
+        return explanation;
+
+    } catch (error) {
+        console.error('Errore:', error);
+        return 'Spiegazione non disponibile'; 
+    }
 };
 
-// Generate AI response based on retrieved chunks
-const generateDocumentBasedResponse = async (searchPhrase, documentContents, followUp = false, followUpText = '') => {
-  try {
-    const formattedDocs = documentContents
-      .map((doc, index) => `DOCUMENT ${index + 1}: ${doc.title}\n${doc.content}`)
-      .join('\n\n---\n\n');
-
-    let prompt;
-    if (followUp && lastExplanation) {
-      prompt = `Sei un assistente specializzato nella ricerca di documenti. Usa ESCLUSIVAMENTE le informazioni nei documenti forniti per rispondere.
-
-Domanda originale: "${lastSearchPhrase}"
-Risposta precedente: "${lastExplanation}"
-Domanda di follow-up: "${followUpText}"
-
-Documenti rilevanti:
-${formattedDocs}
-
-Rispondi alla domanda di follow-up basandoti solo sui documenti forniti. Se l'informazione specifica non è presente, usa i contenuti disponibili per dedurre una risposta ragionevole, citando i documenti rilevanti. Se i documenti non contengono nulla di pertinente, rispondi: "Mi dispiace, questa informazione non è disponibile nei documenti che ho analizzato."`;
-    } else {
-      prompt = `Sei un assistente specializzato nella ricerca di documenti. Usa ESCLUSIVAMENTE le informazioni nei documenti forniti per rispondere.
-
-Domanda: "${searchPhrase}"
-
-Documenti rilevanti:
-${formattedDocs}
-
-Rispondi alla domanda basandoti solo sui documenti forniti. Se l'informazione specifica non è presente, usa i contenuti disponibili per dedurre una risposta ragionevole, citando i documenti rilevanti (includendo titolo e contenuto). Se i documenti non contengono nulla di pertinente, rispondi: "Mi dispiace, questa informazione non è disponibile nei documenti che ho analizzato."`;
-    }
-
-    console.log('Prompt:', prompt);
-    const response = await openaiClient.chat.completions.create({
-      model: 'gpt-4o-mini',
-      messages: [{ role: 'user', content: prompt }],
-      temperature: 0.2,
-      max_tokens: 800,
-      top_p: 0.95,
-      frequency_penalty: 0.0,
-      presence_penalty: 0.0,
+const maxDepth = 1; // Imposta la profondità massima desiderata
+function isWithinDepth(filePath, basePath, maxDepth) {
+    const relativePath = path.relative(basePath, filePath);
+    const depth = relativePath.split(path.sep).length - 1; // Calcola la profondità
+    return depth <= maxDepth; // Verifica se la profondità è accettabile
+}
+// Funzione per osservare la cartella e avvisare quando viene aggiunto un documento
+const watchDocumentsFolder = (folderPath) => {
+    const watcher = chokidar.watch(folderPath, {
+        persistent: true,
+        ignored: /(^|[\/\\])\../,  // ignora file nascosti
+        ignoreInitial: true // Ignora gli eventi all'inizio per evitare letture duplicate
     });
-    const explanation = response.choices[0].message.content.trim();
 
-    if (!followUp) {
-      lastSearchPhrase = searchPhrase;
-      lastExplanation = explanation;
-    }
+    // Evento quando un file viene aggiunto
+    watcher.on('add', (filePath) => {
+        if (path.extname(filePath) === '.docx') {
+            //console.log(`Nuovo documento aggiunto: ${filePath}`);
+            initIndex();
 
-    return explanation;
-  } catch (error) {
-    console.error('Error generating response:', error);
-    return 'Mi dispiace, non è stato possibile generare una risposta basata sui documenti.';
-  }
+        } else {
+            console.log(`File "${filePath}" non è un documento DOCX, ignorato.`);
+        }
+    });
+
+    // Eventi di rimozione di un file
+    watcher.on('unlink', (filePath) => {
+        if (isWithinDepth(filePath, folderPath, maxDepth)) {
+            console.log(`File rimosso: ${filePath}`);
+            initIndex();
+        }
+    });
+
+    // Eventi di modifica di un file
+    watcher.on('change', (filePath) => {
+        if (isWithinDepth(filePath, folderPath, maxDepth)) {
+            console.log(`File modificato: ${filePath}`);
+            initIndex();
+        }
+    });
+
+    watcher.on('error', (error) => console.error(`Errore nel watcher: ${error}`));
 };
 
-// Search API endpoint
-app.post('/search_in_doc', async (req, res) => {
-  const { query, limit = 5, followUp = false, followUpText = '' } = req.body;
-  try {
-    if (!followUp) {
-      const searchResults = await searchDocuments(query, limit);
-      if (searchResults.results.length > 0) {
-        const explanation = await generateDocumentBasedResponse(query, lastNDocuments);
-        res.json({ ...searchResults, explanation });
-      } else {
-        res.json({
-          results: [],
-          totalHits: 0,
-          explanation: 'Nessuna corrispondenza trovata nei documenti disponibili.',
-        });
-      }
-    } else {
-      if (lastNDocuments.length === 0) {
-        res.json({
-          results: [],
-          explanation: 'Non ci sono documenti contestuali disponibili. Per favore esegui una nuova ricerca.',
-        });
-        return;
-      }
-      const explanation = await generateDocumentBasedResponse(
-        query || lastSearchPhrase,
-        lastNDocuments,
-        true,
-        followUpText
-      );
-      res.json({ results: [], explanation });
+const initIndex = async () => {
+    indexName = `${baseIndexName}_${Date.now()}`;
+    await createIndex();
+
+    for (const folderPath of folderPaths) {
+        await readDocumentsFromFolder(folderPath);
+        watchDocumentsFolder(folderPath);
     }
-  } catch (error) {
-    console.error('Search error:', error);
-    res.status(500).json({ error: 'Errore durante la ricerca', message: error.message });
-  }
-});
-
-// Document folder configuration
-const folderPaths = ['./private/Editoriale', './private/Alert Normativo'];
-const maxDepth = 2;
-let baseFolderDepth = 0;
-
-// Check depth for directory traversal
-function isWithinDepth(filePath, basePath) {
-  const relativePath = path.relative(basePath, filePath);
-  const depth = relativePath.split(path.sep).length;
-  return depth <= maxDepth;
 }
 
-// Watch folder for changes
-const watchDocumentsFolder = (folderPath) => {
-  const watcher = chokidar.watch(folderPath, {
-    persistent: true,
-    ignored: /(^|[\/\\])\../,
-    ignoreInitial: true,
-    depth: maxDepth,
-  });
+const folderPaths = ['./private/Editoriale', './private/Alert Normativo'];
 
-  watcher
-    .on('add', async (filePath) => {
-      if (path.extname(filePath).toLowerCase() === '.docx' && isWithinDepth(filePath, folderPath)) {
-        console.log(`New document added: ${filePath}`);
-        await indexDocument(filePath);
-      }
-    })
-    .on('change', async (filePath) => {
-      if (path.extname(filePath).toLowerCase() === '.docx' && isWithinDepth(filePath, folderPath)) {
-        console.log(`Document changed: ${filePath}`);
-        await deleteDocumentVectors(filePath);
-        await indexDocument(filePath);
-      }
-    })
-    .on('unlink', async (filePath) => {
-      if (path.extname(filePath).toLowerCase() === '.docx' && isWithinDepth(filePath, folderPath)) {
-        console.log(`Document removed: ${filePath}`);
-        await deleteDocumentVectors(filePath);
-      }
-    })
-    .on('error', (error) => console.error(`Watcher error: ${error}`));
-};
-
-// Initialize indexing and watchers
-const initIndex = async () => {
-  for (const folderPath of folderPaths) {
-    baseFolderDepth = folderPath.split(path.sep).length;
-    await readDocumentsFromFolder(folderPath);
-    watchDocumentsFolder(folderPath);
-  }
-  console.log(`Pinecone index "${indexName}" initialized and ready for searching.`);
-};
-
-// Health check endpoint
-app.get('/health', (req, res) => {
-  res.json({ status: 'OK', indexName });
-});
-
-// Start server
+// Initialize Index and Start Server
 const startServer = async () => {
   try {
-    await initIndex();
+    await initIndex(); 
     app.listen(port, () => console.log(`Server is listening on port ${port}`));
   } catch (error) {
-    console.error('Failed to initialize index or start server:', error);
-    process.exit(1);
+    console.error("Failed to initialize index or start server:", error);
   }
 };
 
+// Entry point
 startServer();
